InsSubNum all wrapped up.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@4830 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ UnitTests/PatPlanT.cs
@@ -5,7 +5,7 @@ using OpenDentBusiness;
 
 namespace UnitTests {
 	public class PatPlanT {
-		public static PatPlan CreatePatPlan(int ordinal,long patNum,long planNum,long subNum){
+		public static PatPlan CreatePatPlan(byte ordinal,long patNum,long planNum,long subNum){
 			PatPlan patPlan=new PatPlan();
 			patPlan.Ordinal=ordinal;
 			patPlan.PatNum=patNum;
